Add generateCode util
