+++ src/app/api/auth/route.ts
@@ -1,7 +1,6 @@
-import { createHash } from "crypto"
-import { NextResponse } from "next/server"
 import { db } from "@/firebase/firebase-app"
 import { doc, serverTimestamp, updateDoc } from "firebase/firestore"
+import { NextResponse } from "next/server"
 import qs from "qs"
 
 const MYANIMELIST_CLIENT_ID = process.env.X_MAL_CLIENT_ID + ""
@@ -80,30 +79,4 @@ export async function GET(request: Request, { params }: { params: any }) {
 
 export const getServerSideProps = ({ req, res }: { req: any; res: any }) => {
   return { props: {} }
-}
-
-export function generateCodeVerifier() {
-  const codeVerifier = generateRandomString(128)
-  return codeVerifier
-}
-
-export function generateCodeChallenge(codeVerifier: string) {
-  const hashed = createHash("sha256")
-    .update(codeVerifier)
-    .digest("base64")
-    .replace(/\+/g, "-")
-    .replace(/\//g, "_")
-    .replace(/=/g, "")
-  return hashed
-}
-
-function generateRandomString(length: number) {
-  let text = ""
-  const possible = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-._~"
-
-  for (let i = 0; i < length; i++) {
-    text += possible.charAt(Math.floor(Math.random() * possible.length))
-  }
-
-  return text
 }
